refactor: rename classes and shift them between packages
